requirements.txt
asyncpg==0.30.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.3.0
